refactor(plot): move get_callback_param_desc into util.py
